Add bilstm tests and more documentation
