+++ docs/source/medacy.nn.rst
@@ -1,0 +1,10 @@
+MedaCy NN
+===========================================
+
+BiLSTM-CRF
+------------------
+
+.. autoclass:: medacy.nn.BiLstmCrf()
+    :members:
+    :undoc-members:
+    :show-inheritance:
