+++ docs/source/medacy.pipeline_components.learners.rst
@@ -1,0 +1,10 @@
+Pipeline Components: Learners
+===========================================
+
+BiLSTM-CRF Learner
+------------------
+
+.. autoclass:: medacy.pipeline_components.BiLstmCrfLearner()
+    :members:
+    :undoc-members:
+    :show-inheritance:
