+++ docs/source/medacy.rst
@@ -5,6 +5,7 @@ medacy package
     medacy.__main__
     medacy.data
     medacy.ner
+    medacy.nn
     medacy.relation
     medacy.pipeline_components
     medacy.tools
