--- medacy/ner/learners/__init__.py
@@ -1,1 +1,0 @@
-from .bilstm_crf import BiLstmCrfLearner
