+++ medacy/ner/pipelines/__init__.py
@@ -4,5 +4,5 @@ from .fda_nano_drug_label_pipeline import FDANanoDrugLabelPipeline
 from .scispacy_pipeline import ScispacyPipeline
 from .drug_event_pipeline import DrugEventPipeline
 from .testing_pipeline import TestingPipeline
-from .lstm_clinical_pipeline import LstmClinicalPipeline
+from .lstm_systematic_review_pipeline import LstmSystematicReviewPipeline
 from .base.base_pipeline import BasePipeline
