+++ medacy/nn/__init__.py
@@ -1,0 +1,2 @@
+"""medacy.nn"""
+from .bilstm_crf import BiLstmCrf
