+++ medacy/pipeline_components/__init__.py
@@ -1,16 +1,22 @@
+"""Import pipeline components"""
+from .annotation.gold_annotator_component import GoldAnnotatorComponent
+from .annotation.biluo_annotator_component import BiluoAnnotatorComponent
+
+from .feature_extraction.discrete_feature_extractor import FeatureExtractor
+from .feature_extraction.word_only_feature_extractor import WordOnlyFeatureExtractor
+
+from .learners.bilstm_crf_learner import BiLstmCrfLearner
+
+from .lexicon import LexiconComponent
+
 from .metamap.metamap import MetaMap
+from .metamap.metamap_component import MetaMapComponent
+
+from .patterns import TableMatcherComponent
 
 from .tokenization.clinical_tokenizer import ClinicalTokenizer
 from .tokenization.character_tokenizer import CharacterTokenizer
 from .tokenization.systematic_review_tokenizer import SystematicReviewTokenizer
-
-from .metamap.metamap_component import MetaMapComponent
-from .annotation.gold_annotator_component import GoldAnnotatorComponent
-from .annotation.biluo_annotator_component import BiluoAnnotatorComponent
-
-from .lexicon import LexiconComponent
-
-from .patterns import TableMatcherComponent
 
 from .units.unit_component import UnitComponent
 from .units.mass_unit_component import MassUnitComponent
@@ -18,7 +24,3 @@ from .units.volume_unit_component import VolumeUnitComponent
 from .units.time_unit_component import TimeUnitComponent
 from .units.frequency_unit_component import FrequencyUnitComponent
 from .units.measurement_unit_component import MeasurementUnitComponent
-
-
-from .feature_extraction.discrete_feature_extractor import FeatureExtractor
-from .feature_extraction.word_only_feature_extractor import WordOnlyFeatureExtractor
